requirements.txt
lightfm==1.17
numpy==1.22.4
pandas==1.4.0
Pillow==10.1.0
plotly==5.10.0
pycountry==22.3.5
requests==2.28.1
streamlit==1.28.1
